+++ source/renderer/app/components/staking/navigation/StakingNavigation.js
@@ -39,7 +39,6 @@ type Props = {
   onNavItemClick: Function,
   isActiveNavItem: Function,
   isIncentivizedTestnet: boolean,
-  isShelleyTestnet: boolean,
 };
 
 @observer
@@ -54,7 +53,6 @@ export default class StakingNavigation extends Component<Props> {
       activeItem,
       isActiveNavItem,
       isIncentivizedTestnet,
-      isShelleyTestnet,
     } = this.props;
     const { intl } = this.context;
     const navigationItems = [
@@ -71,13 +69,11 @@ export default class StakingNavigation extends Component<Props> {
         label: intl.formatMessage(messages.rewards),
       },
     ];
-    if (!isIncentivizedTestnet && !isShelleyTestnet) {
-      navigationItems.push({
-        id: 'epochs',
-        label: intl.formatMessage(messages.epochs),
-      });
-    }
-    if (isShelleyTestnet) {
+    // navigationItems.push({
+    //   id: 'epochs',
+    //   label: intl.formatMessage(messages.epochs),
+    // });
+    if (!isIncentivizedTestnet) {
       navigationItems.push({
         id: 'info',
         label: intl.formatMessage(messages.info),